+++ src/RuleEnrolment/Log/Repository.php
@@ -22,6 +22,7 @@ final class Repository
 
     use DICTrait;
     use SrUserEnrolmentTrait;
+
     const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
     /**
      * @var self|null
@@ -72,7 +73,7 @@ final class Repository
      */
     public function deleteUserLogs(int $user_id)/*: void*/
     {
-        foreach ($this->getLogs(null, null, null, null, null, null, null, null, null, $user_id) as $log) {
+        foreach ($this->getLogs(null, null, null, null, null, null, null, null, null, $user_id) + $this->getLogs(null, null, null, null, null, null, null, null, null, null, $user_id) as $log) {
             $this->deleteLog($log);
         }
     }
@@ -123,16 +124,17 @@ final class Repository
      * @param ilDateTime|null $date_end
      * @param int|null        $status
      * @param int|null        $user_id
+     * @param int|null        $execute_user_id
      *
      * @return Log[]
      */
     public function getLogs(/*?*/ int $object_id = null, /*?*/ string $sort_by = null, /*?*/ string $sort_by_direction = null, /*?*/ int $limit_start = null, /*?*/ int $limit_end = null, /*?*/
-        string $message = null, /*?*/ ilDateTime $date_start = null, /*?*/ ilDateTime $date_end = null, /*?*/ int $status = null, /*?*/ int $user_id = null
+        string $message = null, /*?*/ ilDateTime $date_start = null, /*?*/ ilDateTime $date_end = null, /*?*/ int $status = null, /*?*/ int $user_id = null,/*?*/ int $execute_user_id = null
     ) : array {
 
         $sql = 'SELECT *';
 
-        $sql .= $this->getLogsQuery($object_id, $sort_by, $sort_by_direction, $limit_start, $limit_end, $message, $date_start, $date_end, $status, $user_id);
+        $sql .= $this->getLogsQuery($object_id, $sort_by, $sort_by_direction, $limit_start, $limit_end, $message, $date_start, $date_end, $status, $user_id, $execute_user_id);
 
         /**
          * @var Log[] $logs
@@ -150,16 +152,17 @@ final class Repository
      * @param ilDateTime|null $date_end
      * @param int|null        $status
      * @param int|null        $user_id
+     * @param int|null        $execute_user_id
      *
      * @return int
      */
     public function getLogsCount(/*?*/ int $object_id = null, /*?*/ string $message = null, /*?*/ ilDateTime $date_start = null, /*?*/ ilDateTime $date_end = null, /*?*/ int $status = null,/*?*/
-        int $user_id = null
+        int $user_id = null,/*?*/ int $execute_user_id = null
     ) : int {
 
         $sql = 'SELECT COUNT(log_id) AS count';
 
-        $sql .= $this->getLogsQuery($object_id, null, null, null, null, $message, $date_start, $date_end, $status, $user_id);
+        $sql .= $this->getLogsQuery($object_id, null, null, null, null, $message, $date_start, $date_end, $status, $user_id, $execute_user_id);
 
         $result = self::dic()->database()->query($sql);
 
@@ -182,11 +185,12 @@ final class Repository
      * @param ilDateTime|null $date_end
      * @param int|null        $status
      * @param int|null        $user_id
+     * @param int|null        $execute_user_id
      *
      * @return string
      */
     private function getLogsQuery(/*?*/ int $object_id = null, /*?*/ string $sort_by = null, /*?*/ string $sort_by_direction = null, /*?*/ int $limit_start = null, /*?*/ int $limit_end = null, /*?*/
-        string $message = null, /*?*/ ilDateTime $date_start = null, /*?*/ ilDateTime $date_end = null, /*?*/ int $status = null,/*?*/ int $user_id = null
+        string $message = null, /*?*/ ilDateTime $date_start = null, /*?*/ ilDateTime $date_end = null, /*?*/ int $status = null,/*?*/ int $user_id = null,/*?*/ int $execute_user_id = null
     ) : string {
 
         $sql = ' FROM ' . self::dic()->database()->quoteIdentifier(Log::TABLE_NAME);
@@ -215,6 +219,10 @@ final class Repository
 
         if (!empty($user_id)) {
             $wheres[] = 'user_id=' . self::dic()->database()->quote($user_id, ilDBConstants::T_INTEGER);
+        }
+
+        if (!empty($execute_user_id)) {
+            $wheres[] = 'execute_user_id=' . self::dic()->database()->quote($execute_user_id, ilDBConstants::T_INTEGER);
         }
 
         if (count($wheres) > 0) {
@@ -296,15 +304,17 @@ final class Repository
 
         if (empty($log->getLogId())) {
             $log->withDate($date);
+            $log->withExecuteUserId(self::dic()->user()->getId());
         }
 
         $log->withLogId(self::dic()->database()->store(Log::TABLE_NAME, [
-            "object_id" => [ilDBConstants::T_INTEGER, $log->getObjectId()],
-            "rule_id"   => [ilDBConstants::T_TEXT, $log->getRuleId()],
-            "user_id"   => [ilDBConstants::T_INTEGER, $log->getUserId()],
-            "date"      => [ilDBConstants::T_TEXT, $log->getDate()->get(IL_CAL_DATETIME)],
-            "status"    => [ilDBConstants::T_INTEGER, $log->getStatus()],
-            "message"   => [ilDBConstants::T_TEXT, $log->getMessage()]
+            "object_id"       => [ilDBConstants::T_INTEGER, $log->getObjectId()],
+            "rule_id"         => [ilDBConstants::T_TEXT, $log->getRuleId()],
+            "user_id"         => [ilDBConstants::T_INTEGER, $log->getUserId()],
+            "execute_user_id" => [ilDBConstants::T_INTEGER, $log->getExecuteUserId()],
+            "date"            => [ilDBConstants::T_TEXT, $log->getDate()->get(IL_CAL_DATETIME)],
+            "status"          => [ilDBConstants::T_INTEGER, $log->getStatus()],
+            "message"         => [ilDBConstants::T_TEXT, $log->getMessage()]
         ], "log_id", $log->getLogId()));
 
         $this->keepLog($log);