+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [3.12.0]
+- Excel import
+  - Create users direct in a global or local roles
+  - Log user which executed
+
 ## [3.11.2]
 - Assign global roles